// server/src/shared/interfaces/csv-row.interface.ts
export interface CsvRow {
  agent: string;
  userType: string;
  policy_mode: string;
  producer: string;
  policy_number: string;
  premium_amount_written: string;
  premium_amount: string;
  policy_type: string;
  company_name: string;
  category_name: string;
  policy_start_date: string;
  policy_end_date: string;
  csr: string;
  account_name: string;
  email: string;
  gender: string;
  firstname: string;
  city: string;
  account_type: string;
  phone: string;
  address: string;
  state: string;
  zip: string;
  dob: string;

}